README: clarify differences between key and token
This commit makes some changes in the installation docs so it's clear that
both the API token and the access key are requisites and that it's clear what
are the differences between them.

I've also updated the configuration files that I had under `deploy/` because I
was using the same value ofr both the key and the token, very confusing!